chore: move unused scripts outside `src`
