+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@interlay/interbtc-api",
-  "version": "1.5.14",
+  "version": "1.5.15",
   "description": "JavaScript library to interact with interBTC and kBTC",
   "main": "build/src/index.js",
   "typings": "build/src/index.d.ts",
@@ -27,8 +27,8 @@
     "generate:defs": "ts-node --skip-project node_modules/.bin/polkadot-types-from-defs --package @interlay/interbtc/interfaces --input ./src/interfaces --endpoint ./src/json/parachain.json",
     "generate:meta": "ts-node --skip-project node_modules/.bin/polkadot-types-from-chain --package @interlay/interbtc/interfaces --endpoint ./src/json/parachain.json --output ./src/interfaces",
     "initialize": "run-s build && ts-node src/utils/setup --initialize",
-    "hrmp-setup": "ts-node src/utils/hrmp-setup",
-    "xcm-cross-chain-transfer": "ts-node src/utils/xcm-cross-chain-transfer",
+    "hrmp-setup": "ts-node scripts/hrmp-setup",
+    "xcm-cross-chain-transfer": "ts-node scripts/xcm-cross-chain-transfer",
     "test": "run-s build test:*",
     "test:lint": "eslint src --ext .ts",
     "test:unit": "mocha test/unit/*.test.ts test/unit/**/*.test.ts",
