+++ tsconfig.json
@@ -35,9 +35,6 @@
     ],
     "exclude": [
         "./node_modules/*",
-        "./src/utils/hrmp-setup.ts",
-        "./src/utils/xcm-cross-chain-transfer.ts",
-        "./src/utils/setup.ts",
-
+        "scripts/*"
     ]
 }
